requirements.txt
amqp==5.3.1
asgiref==3.9.1
billiard==4.2.1
black==25.1.0
celery==5.5.3
cfgv==3.4.0
click==8.2.1
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
coverage==7.10.2
distlib==0.4.0
Django==5.2.5
django-jazzmin==3.0.1
django-tenants==3.8.0
djangorestframework==3.16.0
filelock==3.18.0
flake8==7.3.0
gunicorn
identify==2.6.12
kombu==5.5.4
mccabe==0.7.0
mypy_extensions==1.1.0
nodeenv==1.9.1
packaging==25.0
pathspec==0.12.1
platformdirs==4.3.8
pillow
pre_commit==4.2.0
prompt_toolkit==3.0.51
psycopg2-binary==2.9.10
pycodestyle==2.14.0
pyflakes==3.4.0
python-dateutil==2.9.0.post0
PyYAML==6.0.2
redis==6.3.0
six==1.17.0
sqlparse==0.5.3
tzdata==2025.2
vine==5.1.0
virtualenv==20.33.1
wcwidth==0.2.13
# For GCS file storage
django-storages[gcs]
# For fetching secrets from Secret Manager
google-cloud-secret-manager